chore(gazette): drop changes that duplicate #2074
#2074 already covers a subset of the items addressed here. Removing the
duplicates to keep the PRs from racing on the same files:

- Revert: move of getSearchSGAuthToken + pushDocumentsForIngestion (D3).
  #2074 relocates these into gazette.service.ts and exports the related
  constants from searchsg.service.ts — different module than the
  reviewer asked for, but that decision belongs in the #2074 review,
  not here.
- Revert: cron test file changes (parseFullTextFromPDF spy module fix,
  matchObject assertion update). #2074 carries the equivalent fix via
  the seed helper.
- Re-align .env.test additions to use #2074's exact values
  (very-cool-index / cool-bucket / cool-domain) so the change is
  textually identical and merges cleanly.

After this commit #2200 carries: defence-in-depth siteId scoping,
audit-delta fix, preserve deletedAt, schema cleanups + fileKey
traversal guard, hook schema import + shared handleAssetUpload,
getBlob named args + inline addDays, new router/service tests, and
the required_error -> Zod 4 error syntax fix.

Note: typecheck on this branch is in the same partial state as #2027's
base for the cron test/prod files — those errors are inherited and
resolved by #2074.
