+++ apps/studio/.env.test
@@ -25,6 +25,6 @@ GROWTHBOOK_CLIENT_KEY=sdk-lnL5Et90CD9sC13e
 NEXT_PUBLIC_GROWTHBOOK_CLIENT_KEY=sdk-lnL5Et90CD9sC13e
 
 SEARCHSG_API_KEY=test-searchsg-api-key
-S3_GAZETTE_BUCKET_NAME=test-gazette-bucket
-EGAZETTE_DOCUMENT_INDEX=test-index
-SEARCHSG_API_KEY=test-key
+EGAZETTE_DOCUMENT_INDEX="very-cool-index"
+S3_GAZETTE_BUCKET_NAME="cool-bucket"
+S3_GAZETTE_DOMAIN_NAME="cool-domain"
