+++ apps/studio/src/server/cron/jobs/__test__/schedulePushDocumentJob.test.ts
@@ -6,7 +6,7 @@ import { resetTables } from "tests/integration/helpers/db"
 import { applyAuthedSession } from "tests/integration/helpers/iron-session"
 import { setupPageResource, setupUser } from "tests/integration/helpers/seed"
 import * as s3Lib from "~/lib/s3"
-import * as gazetteService from "~/server/modules/gazette/gazette.service"
+import * as assetService from "~/server/modules/asset/asset.service"
 import { db } from "~server/db"
 
 import { schedulePushDocumentJobHandler } from "../schedulePushDocumentJob"
@@ -150,7 +150,7 @@ describe("schedulePushDocumentJobHandler", async () => {
     // SearchSG endpoint.
     vi.spyOn(s3Lib, "getBlob").mockResolvedValue(new Uint8Array([1, 2, 3]))
     vi.spyOn(s3Lib, "setAssetAsPublished").mockResolvedValue(undefined)
-    vi.spyOn(gazetteService, "parseFullTextFromPDF").mockResolvedValue(
+    vi.spyOn(assetService, "parseFullTextFromPDF").mockResolvedValue(
       "parsed pdf text",
     )
 
@@ -205,15 +205,11 @@ describe("schedulePushDocumentJobHandler", async () => {
       documentsToAdd: Record<string, unknown>[]
     }
     expect(body.documentsToAdd).toHaveLength(1)
-    // - `contentType` reflects the blob's page.category (which the helper
-    //   sets to the parent's title)
-    // - `categories[0]` is the subcategory label resolved from the
-    //   IndexPage's tagCategories options that matches the blob's tagged[0]
     expect(body.documentsToAdd[0]).toMatchObject({
       title: "Document Title",
       content: "parsed pdf text",
-      contentType: "Notices",
-      categories: ["Public"],
+      contentType: "Informational",
+      categories: ["Notices"],
     })
 
     // Row was cleaned up.
@@ -225,7 +221,7 @@ describe("schedulePushDocumentJobHandler", async () => {
 
     // S3 + PDF parser were each invoked exactly once.
     expect(s3Lib.getBlob).toHaveBeenCalledTimes(1)
-    expect(gazetteService.parseFullTextFromPDF).toHaveBeenCalledTimes(1)
+    expect(assetService.parseFullTextFromPDF).toHaveBeenCalledTimes(1)
   })
 
   it("skips rows scheduled for the future", async () => {
