+++ apps/studio/src/server/modules/searchsg/searchsg.service.ts
@@ -162,71 +162,3 @@ interface SearchSGConfig {
     }
   }
 }
-
-export interface PushDocument {
-  documentId: string
-  title: string
-  url: string
-  contentType: string
-  content: string
-  date: string
-  categories: (string | undefined)[]
-}
-
-const getSearchSGAuthToken = async () => {
-  const response = await fetch(`${SEARCHSG_BASE_URL}${SearchSgApi.Auth}`, {
-    method: "POST",
-    headers: {
-      Authorization: `Basic ${env.SEARCHSG_API_KEY}`,
-      "User-Agent": ISOMER_UA,
-    },
-  })
-
-  if (!response.ok) {
-    throw new Error(`Failed to get SearchSG auth token: ${response.statusText}`)
-  }
-
-  const { accessToken, tokenType } = (await response.json()) as {
-    accessToken: string
-    tokenType: string
-  }
-
-  return { accessToken, tokenType }
-}
-
-export const pushDocumentsForIngestion = async (documents: PushDocument[]) => {
-  if (documents.length === 0) {
-    return
-  }
-
-  const { accessToken, tokenType } = await getSearchSGAuthToken()
-
-  const response = await fetch(
-    `${SEARCHSG_BASE_URL}/v2/indexes/${env.EGAZETTE_DOCUMENT_INDEX}/documents`,
-    {
-      method: "POST",
-      headers: {
-        Authorization: `${tokenType} ${accessToken}`,
-        "Content-Type": "application/json",
-        "User-Agent": ISOMER_UA,
-      },
-      body: JSON.stringify({ documentsToAdd: documents }),
-    },
-  )
-
-  if (!response.ok) {
-    const errorText = await response.text()
-    logger.error(
-      { status: response.status, error: errorText },
-      "Failed to push documents for ingestion",
-    )
-    throw new Error(
-      `Failed to push documents for ingestion: ${response.statusText}`,
-    )
-  }
-
-  logger.info(
-    { count: documents.length },
-    "Successfully pushed documents for ingestion",
-  )
-}
